Fuel is nolonger expect/actual class.
