+++ fuel/src/appleMain/kotlin/fuel/FuelBuilder.kt
@@ -7,10 +7,6 @@ public actual class FuelBuilder {
 
     private val defaultSessionConfiguration by lazy { NSURLSessionConfiguration.defaultSessionConfiguration }
 
-    init {
-        failOnNativeIfLegacyMemoryManager()
-    }
-
     public fun config(sessionConfiguration: NSURLSessionConfiguration): FuelBuilder = apply {
         this.sessionConfiguration = sessionConfiguration
     }
