+++ fuel/src/commonMain/kotlin/fuel/FuelBuilder.kt
@@ -1,5 +1,5 @@
 package fuel
 
-public expect class FuelBuilder {
+public expect class FuelBuilder() {
     public fun build(): HttpLoader
 }
